refactor(pgp): fold uselesskey-pgp-native content into uselesskey-pgp::native
Inverts the dep direction between `uselesskey-pgp` and
`uselesskey-pgp-native`. The `PgpNativeExt` trait and its impls now
live in `crates/uselesskey-pgp/src/native.rs`, gated behind a new
`native` Cargo feature on `uselesskey-pgp`. `uselesskey-pgp-native`
becomes a thin re-export shim re-exporting `uselesskey_pgp::native::*`
to preserve the v0.7.x public-internal surface for downstream pinners.

This matches the v0.7.0 fold pattern established for `core-cache`,
`core-id`, `core-hash`, `core-factory`, `token-spec`, etc. and the
v0.8.0 hmac-spec fold (#595) where the owner fixture crate took
canonical ownership and the previously-published microcrate became a
compatibility shim.

Mechanical changes:

- New `crates/uselesskey-pgp/src/native.rs` carries `PgpNativeExt`,
  the impl block on `PgpKeyPair`, and the inline `#[cfg(test)] mod
  tests` block.
- `crates/uselesskey-pgp/src/lib.rs` declares
  `#[cfg(feature = "native")] pub mod native;` and re-exports
  `PgpNativeExt` at the top level under the same gate.
- `uselesskey-pgp/Cargo.toml` adds a `native` Cargo feature
  (`native = []`); the `pgp` dep is already a required workspace
  dep for keygen, so no new dep surface is added. The `pgp` 0.19
  pin moves into `[workspace.dependencies]` so the two crates share
  the same version line.
- `uselesskey-pgp-native/Cargo.toml` drops its direct `pgp = ...`
  dep, adds `uselesskey-pgp.workspace = true` with
  `features = ["native"]`, and routes its `pgp-native` feature to a
  no-op marker (the surface is unconditionally re-exported because
  `uselesskey-pgp` is now a hard dep with `native` already enabled).
  `pgp` remains as a dev-dep for the integration test and example.
- `uselesskey-pgp-native/src/lib.rs` is now a thin shim:
  `pub use uselesskey_pgp::native::*;`.
- `xtask::PUBLISH_CRATES` moves `uselesskey-pgp-native` to a
  post-`uselesskey-pgp` slot under a "PGP compatibility shim"
  comment, matching `uselesskey-core-hmac-spec` after the hmac fold.
- `xtask::plan::dependents()` makes `uselesskey-pgp-native` a leaf
  and adds it as a `uselesskey-pgp` dependent (so a pgp source
  change still triggers a shim rebuild). A new
  `pgp_native_change_is_isolated_to_shim` test mirrors the
  `core_hmac_spec_change_is_isolated_to_shim` shape.
- `docs/explanation/architecture.md`,
  `docs/metadata/workspace-docs.json`, and
  `docs/reference/support-matrix.md` describe the crate as a
  compatibility shim.

Both `uselesskey-pgp` and `uselesskey-pgp-native` bump to 0.7.2 so
the shim's re-export of `uselesskey_pgp::native` resolves against a
registry version that ships the new path: `cargo publish --dry-run
-p uselesskey-pgp-native` would otherwise resolve against the
published v0.7.1 of `uselesskey-pgp`, which does not have `native`,
and fail with `E0432: unresolved imports`.

Downstream consumers (`uselesskey`, `uselesskey-interop-tests`) keep
their existing `version = "0.7.1"` constraint — cargo treats it as
`^0.7.1`, which accepts v0.7.2.

This is PR-3 of the v0.8.0 SRP collapse lane. The shim crate stays
published for v0.7.x compat and is scheduled for removal in PR-4.
