+++ crates/uselesskey-pgp-native/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
 name = "uselesskey-pgp-native"
-version = "0.7.1"
+version = "0.7.2"
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 repository.workspace = true
-description = "Native OpenPGP key type adapters for uselesskey fixtures."
+description = "Deprecated compatibility shim re-exporting PgpNativeExt from uselesskey-pgp."
 categories.workspace = true
 keywords = ["openpgp", "pgp", "fixtures", "testing", "adapter"]
 readme = "README.md"
@@ -15,19 +15,18 @@ documentation = "https://docs.rs/uselesskey-pgp-native"
 authors.workspace = true
 
 [dependencies]
-uselesskey-pgp = { path = "../uselesskey-pgp", version = "0.7.1", optional = true }
-pgp = { version = "0.19.0", default-features = false }
+uselesskey-pgp = { workspace = true, features = ["native"] }
 
 [package.metadata.docs.rs]
 features = ["pgp-native"]
 
 [features]
 default = ["pgp-native"]
-pgp-native = ["dep:uselesskey-pgp"]
+pgp-native = []
 
 [dev-dependencies]
-uselesskey-pgp = { path = "../uselesskey-pgp", version = "0.7.1" }
 uselesskey-core = { path = "../uselesskey-core", version = "0.7.1" }
+pgp = { workspace = true }
 
 
 [lints]
